ingress-api/requirements.txt
fastapi
pydantic
google-cloud-pubsub